Changed slack config port
The default port was a string, which we can't use with a socket. Changed it to an arbitrary number to prevent that issue from occurring, since new users may not realize that this needs to be a number. I spent a second trying to sort out why "8085" wasn't working before I realized string was no good.