image screen shot plugin phantomjs process fix
Noticed that with each call to this command it was spawning a new instance of the phantomjs process and not cleaning up after itself properly. Since the code only allowed one request to be running at a time opted to change the code to instantiate the phantomjs process during the initialize call and reference that for every screenshot request rather then making new instances.