Update api_calls with a version-resistant wrapper.
Since
https://github.com/slackhq/python-slackclient/commit/ac343caf6a3fd8f4b16a79246264a05a7d257760
SlackClient.api_call returns a pre-parsed dict instead of a JSON
string.  To provide compatibility to this new version (a week ago)
and retain compatibility backwards, api_call()s in slackrtm.py are
filtered through a duck-hunt wrapper that does whatever is needed
to return a dict, regardless of the initial response type.