Prevent Slack from printing bot messages
SlackRTM, upon receiving a Slack message, prints it out in Hangouts. It then sees that a new message is in Hangouts and sends it back to Slack. It appears this is not happening in the reverse because Slack isn't sending us our message back as a new message however hangups does. This change does have a side effect of not sending any of the bot text printed in Hangouts over to Slack, however, this makes Slack useable and it is worth it to me. 

Also, since my bot does not have a photo, the plugin is logging exceptions frequently. This is not an exceptional state and a user not having a photo_url should be handled prior to an exception.